% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiod.R
\name{aggregates}
\alias{aggregates}
\title{Extract the aggregate variables from a WIOD data.table}
\usage{
aggregates(wiot)
}
\arguments{
\item{wiot}{WIOD data.table returned by \code{\link[wiod.diagrammer]{loadWIOD}}}
}
\value{
A \code{\link[data.table]{data.table}} with 10 columns
(all numeric, in US dollars as in the WIOD source file, except for \code{SectorNr} and\code{Country}):
\describe{
  \item{\code{SectorNr}}{WIOD numeric codes for industries and final use sectors (integer vector)}
  \item{\code{Country}}{Three-letter ISO country code, as in the WIOD file (character vector)}
  \item{\code{II_fob}}{Total intermediate consumption}
  \item{\code{TXSP}}{Taxes less subsidies on products}
  \item{\code{EXP_adj}}{CIF/FOB adjustments on exports}
  \item{\code{PURR}}{Direct purchases abroad by residents}
  \item{\code{PURNR}}{Purchases on the domestic territory by non-residents }
  \item{\code{VA}}{Value added at basic prices}
  \item{\code{IntTTM}}{International Transport Margins}
  \item{\code{GO}}{Output at basic prices}
}
}
\description{
Extract the aggregate variables from a WIOD data.table
}
