% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiod.R
\name{countries}
\alias{countries}
\title{A table of WIOD countries (their 3-letter codes and full names)}
\usage{
countries()
}
\description{
A \code{\link[data.table]{data.table}} with three columns:
\describe{
  \item{\code{Country}}{Three-letter ISO country code, as in the WIOD file (character vector)}
  \item{\code{CountryLab}}{Full name of a country (character vector)}
  \item{\code{isEUmember}}{Is a country a member of the EU (2017)? (logical vector)}
}
}
