% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiod.R
\name{findLinks}
\alias{findLinks}
\title{Find the specific linkages in the WIOD}
\usage{
findLinks(partners, flat_wiod, start_countries, start_sectors,
  ListOfselectionFuns, by = if (partners == "suppliers") c("ImpCountry",
  "ImpSectorNr") else c("ExpCountry", "ExpSectorNr"))
}
\arguments{
\item{partners}{String: either \code{"suppliers"} or \code{"users"}.}

\item{flat_wiod}{"Flat" WIOD \code{\link[data.table]{data.table}} returned by
\code{\link[wiod.diagrammer]{flatWIOD}}, possibly modified by a user
(e.g. with added columns to be used through \code{by} or with some irrelevant linkages/rows removed).
None of the basic 5 columns returned by \code{\link[wiod.diagrammer]{flatWIOD}}
must be removed.}

\item{start_countries}{Character vector: a three-letter ISO country code,
as in the WIOD file (character vector), see also \code{\link[wiod.diagrammer]{countries}},
of the countries whose top partners are to be found in the first round.}

\item{start_sectors}{Numeric vector: WIOD numeric codes (integers) for industries and
final use sectors, see also \code{\link[wiod.diagrammer]{sectors}},
of the sectors whose top partners are to be found in the first round.}

\item{ListOfselectionFuns}{List (with at least 1 element) of functions, where each
function takes \code{flat_wiod} as an argument and returns a logical vector:
which partners should be selected at each round.}

\item{by}{Optional character vector: the names of the \code{flat_wiod}'s columns containing
the groupings by which the partners will be selected. Default: none i.e.
empty string (\code{""}). Default: importing country and sector (when extracting suppliers)
or exporting country and sector (when extracting customers).}
}
\value{
A \code{\link[data.table]{data.table}} (with added S3 class 'SelectedLinksDT' with
a corresponding \code{\link[wiod.diagrammer]{plot}} implementation) with the same columns
as in \code{flat_wiod} (but ordered differently), with only those rows that represent
the top direct and (possibly) indirect linkages.

This function extracts the first-round (direct) and further (indirect)
backward or forward linkages (i.e. suppliers of suppliers of suppliers etc. or
customers of customers of customers etc.), based on the functions in \code{ListOfselectionFuns}
which take \code{flat_wiod} and for each combination of the dimensions specified in \code{by}
return a logical (Boolean) vector which determines if a given partner is selected or not.

The direction of the flows (backward or forward) is determined by the string argument
\code{partners} which must be either \code{"suppliers"} or \code{"users"}.

The number of rounds is determined by the length of \code{ListOfselectionFuns} i.e.
if this vector contains e.g. two functions, \code{findLinks} will extract the direct linkages and
one layer of indirect linkages. The functions in the \code{breadth_vec} vector determine
which partners selected at round (for each combination/grouping as described above).

The starting point of the search for top partners is determined by the arguments
\code{start_countries} and \code{start_sectors} -- the partners of those countries and
sectors will be searched in the first round.
}
\description{
Find the specific linkages in the WIOD
}
