% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiod.R
\name{flatWIOD}
\alias{flatWIOD}
\title{Transform the WIOD data.table into a flat/long data.table}
\usage{
flatWIOD(wiot, rows = seq_len(nrow(wiot)), columns = seq_len(ncol(wiot)))
}
\arguments{
\item{wiot}{WIOD data.table returned by \code{\link[wiod.diagrammer]{loadWIOD}}}

\item{rows}{\code{wiot}'s rows to be dropped (a numeric vector with negative numbers) or
kept (a numeric vector with positive) prior to flattening/reshaping;
default: all rows kept}

\item{columns}{\code{wiot}'s columns to be dropped (a numeric vector with negative numbers) or
kept (a numeric vector with positive) prior to flattening/reshaping;
default: all columns kept}
}
\value{
A \code{\link[data.table]{data.table}} where each row corresponds to
a combination of a source (exporting) and destination (importing) sectors
and countries, with ten columns:
\describe{
  \item{\code{ExpSectorNr}}{Exporting sector -- WIOD numeric codes for industries (integer vector),
  see also \code{\link[wiod.diagrammer]{sectors}}}
  \item{\code{ExpCountry}}{Exporting country -- a three-letter ISO country code,
  as in the WIOD file (character vector), see also \code{\link[wiod.diagrammer]{countries}}}
  \item{\code{value}}{An absolute value of the flow (in US dollars as in the WIOD source file)}
  \item{\code{ImpSectorNr}}{Importing sector -- WIOD numeric codes for industries and
  final use sectors (integer vector), see also \code{\link[wiod.diagrammer]{sectors}}}
  \item{\code{ImpCountry}}{Importing country -- a three-letter ISO country code,
  as in the WIOD file (character vector), see also \code{\link[wiod.diagrammer]{countries}}}
}
}
\description{
Transform the WIOD data.table into a flat/long data.table
}
