% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wiod.R
\name{loadWIOD}
\alias{loadWIOD}
\title{Load WIOD data from a WIOD .RData file}
\usage{
loadWIOD(file_name = "WIOT2014_October16_ROW.RData")
}
\arguments{
\item{file_name}{A character string giving the name of the file. Default:
\code{"WIOT2014_October16_ROW.RData"}. The file must contain an object
named \code{wiot}.}
}
\value{
The \code{wiot} object contained in the .Rdata file
(a \code{\link[data.table]{data.table}}).
}
\description{
Loads WIOD data for a given year from an official WIOD .Rdata file
(release 2016), which has to be first downloaded from
\url{http://www.wiod.org/protected3/data16/wiot_ROW/wiot_r_Nov16.zip}
and extracted (unzipped).
}

