% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrammer.R
\name{normalise}
\alias{normalise}
\title{Rescale values to the [0,1] range}
\usage{
normalise(numvec)
}
\arguments{
\item{numvec}{A numeric vector}
}
\value{
A numeric vector with values in the range [0,1].
}
\description{
A helper function -- normalise/rescale values so that they fit
into a range from 0 to 1. (The minimum value becomes 0 and the
maximum value becomes 1).
}
\examples{
x <- seq(from=-3.7, to=6.5, by=1.36)
y <- normalise(x)
data.frame(x,y)
#     x          y
# -3.70  0.0000000
# -2.34  0.1428571
# -0.98  0.2857143
#  0.38  0.4285714
#  1.74  0.5714286
#  3.10  0.7142857
#  4.46  0.8571429
#  5.82  1.0000000
}
