% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagrammer.R
\name{plot.SelectedLinksDT}
\alias{plot.SelectedLinksDT}
\title{Plot a diagram (directed graph) of top linkages}
\usage{
plot.SelectedLinksDT(top_links_dt, wiot, units = list(units_suffix = "bn USD",
  mln_USD_exchange_rate = 0.001),
  country_labels_dt = wiod.diagrammer::countries(),
  sector_labels_dt = wiod.diagrammer::sectors(wiot),
  aggregates_dt = wiod.diagrammer::aggregates(wiot),
  arrowSizeFun = function(value) wiod.diagrammer:::normalise(value) * 16 + 1,
  arrowLabelFun = identity, nodeSizeFun = function(country_sector_dt)
  country_sector_dt \%>\% ifelse.(.$isFinal, .$II_fob, .$GO),
  nodeSizeTransformFun = function(node_size_val)
  wiod.diagrammer::normalise(log(node_size_val) + 1) * 8 + 8,
  nodeLabelFun = function(country_sector_dt) country_sector_dt \%>\%
  paste.(.$CountryLab, .$SectorLab, paste(.$NodeSize \%>\% numberFormattingFun,
  units$units_suffix), sep = ", "),
  specificNodeOptionsFun = function(country_sector_dt)
  ifelse(country_sector_dt$isFinal, "fontname=\\"times-italic\\"", ""),
  specificArrowOptionsFun = function(top_links_dt)
  ifelse(top_links_dt$ExpCountry == top_links_dt$ImpCountry, "style=solid",
  "style=dashed"), general_arrow_options = "color=grey",
  general_node_options = "shape=box", graph_options = c("layout=dot",
  "rankdir=LR"), numberFormattingFun = function(value) value \%>\%
  formatC(digits = 1, format = "f", big.mark = " "), nchar_wrap = 30,
  gvcode = FALSE)
}
\arguments{
\item{top_links_dt}{A data.table returned by \code{\link[wiod.diagrammer]{topLinks}}}

\item{wiot}{WIOD data.table returned by \code{\link[wiod.diagrammer]{loadWIOD}}}

\item{units}{Optional: a list of two named interrelated elements:
\describe{
  \item{\code{units_suffix}}{a string (character vector of length 1) with a name
  of the monetary units (default: \code{"bn USD"}). This element
  should be consistent with the next element \code{mln_USD_exchange_rate}}!
  \item{\code{mln_USD_exchange_rate}}{a single value (numeric vector of length 1)
  by which the original WIOD numbers (which are in million US dollars) will be
  multiplied (default: \code{0.001}, so that the values become billion US dollars).
  This element should be consistent with the previous element \code{units_suffix}!}
}}

\item{country_labels_dt}{Optional: a \link[data.table]{data.table} with specific
columns -- see the documentation for \link[wiod.diagrammer]{countries} which is
the function called to produce the default value for \code{country_labels_dt}.
With this data.table, you may provide alternative (non-default) country names/labels
and/or additional country-level variables (columns) to be used by your (i.e. non-default)
functions for the arguments \code{nodeSizeFun}, \code{nodeLabelFun}, and
\code{specificNodeOptionsFun}.}

\item{sector_labels_dt}{Optional: a \link[data.table]{data.table} with specific
columns -- see the documentation for \link[wiod.diagrammer]{sectors} which is
the function called to produce the default value for \code{country_labels_dt}
With this data.table, you may provide alternative (non-default) sector names/labels
and/or additional sector-level variables (columns) to be used by your (i.e. non-default)
functions for the arguments \code{nodeSizeFun}, \code{nodeLabelFun}, and
\code{specificNodeOptionsFun}.}

\item{aggregates_dt}{Optional: a \link[data.table]{data.table} with specific
columns -- see the documentation for \link[wiod.diagrammer]{aggregates} which is
the function called to produce the default value for \code{aggregates_dt}.
With this data.table, you may provide additional or alternative country-and-sector-level
variables (columns) to be used by your (i.e. non-default) functions for the arguments
\code{nodeSizeFun}, \code{nodeLabelFun}, and \code{specificNodeOptionsFun}.}

\item{arrowSizeFun}{Optional: a one-numeric-argument function which translates
\code{top_links_dt}'s \code{value} column into GraphViz's \code{penwidth} values
for arrows (edges), see \url{http://www.graphviz.org/content/attrs#dpenwidth}).}

\item{arrowLabelFun}{Optional: a one-numeric-argument function which translates
\code{top_links_dt}'s \code{value} column (multiplied by \code{units$mln_USD_exchange_rate},
transformed by the \code{numberFormattingFun} function, and with \code{units_suffix}
\link{paste}d) into GraphViz's \code{label}s for arrows (edges). Default: \link{identity}.}

\item{nodeSizeFun}{Optional: a function which takes a country-and-sector level
\link[data.table]{data.table} (including the columns of \code{aggregates_dt}) and returns
a numeric vector representing the sizes of the country-sector nodes (boxes).
Default: the value of 'output at basic prices' (\code{GO}) for the industrial sectors
(i.e. those that produce something -- goods or services)
and 'total intermediate consumption' (\code{II_fob}) for the final use sectors}

\item{nodeSizeTransformFun}{Optional: a one-argument function which takes
the numeric value returned by the function \code{nodeSizeFun} and translates it
into GraphViz's \code{fontsize} values for the country-sector nodes (boxes), see
\url{http://www.graphviz.org/content/attrs#dfontsize}).}

\item{nodeLabelFun}{Optional: a function which takes the country-and-sector-level
\link[data.table]{data.table} -- a merger of \code{aggregates_dt},
\code{country_labels_dt}, and \code{sector_labels_dt} for the country-sector combinations
present in \code{top_links_dt} -- and returns a character vector used as GraphViz's
\code{label}s for nodes (boxes). If the function needs to refer to the node size
(produced by the function \code{nodeSizeFun}) it should refer to data.table's column
\code{NodeSize}. Default: it returns, for each node, the following
components \link{paste}d together: country label, sector label, a value returned by
\code{nodeSizeFun} and the \code{units$units_suffix}.}

\item{specificNodeOptionsFun}{Optional: a function which takes the country-and-sector-level
\link[data.table]{data.table} and returns a character vector of GraphViz's node attributes*,
see ** with 'node' highlighted in
the 'COMPONENTS' section. Default: \code{'fontname="times-italic"'} for the final
use sectors (such as final consumption or investment).}

\item{specificArrowOptionsFun}{Optional: a function which takes \code{top_links_dt}
as an argument and returns a character vector of GraphViz's edge attributes*,
see ** with 'edge' highlighted in
the 'COMPONENTS' section. Default: \code{'style=solid'} for the domestic flows and
'style=dashed' for the international flows.}

\item{general_arrow_options}{Optional: a character vector with the GraphViz's edge
attributes of all the arrows (edges)*. Default: \code{'color=grey'}.}

\item{general_node_options}{Optional: a character vector with the GraphViz's node
attributes of all the nodes (country-sector boxes)*. Default: \code{'shape=box'}.}

\item{graph_options}{Optional: a character with the GraphViz's graph attributes*.
See ** with 'graph' highlighted in the 'COMPONENTS' section.
Default: \code{c('layout=dot', 'rankdir=LR')}.
For the 'layout' see \url{https://en.wikipedia.org/wiki/Graphviz#Software_architecture};
'rankdir' is applicable only if \code{'layout=dot'}, \code{LR} = left-to-right,
\code{TB} = top-to-bottom.}

\item{numberFormattingFun}{Optional: a one-numeric-argument function used to format all
the numbers displayed in the graph. Default: it returns a character vector with values
formated as one digit after decimal dot and blank as a digit-grouping separator.}

\item{nchar_wrap}{Optional: a numeric constant -- the number of characters arounnd which the text
of arrow labels (returned by \code{arrowLabelFun}) is wraped into another line.
Default: 30.}

\item{gvcode}{Optional: should the generated GraphViz code be returned instead
of the \code{\link[DiagrammeR]{grViz}} object (logical)? This is useful for debugging.
Default: FALSE.}
}
\value{
The return value of \code{\link[DiagrammeR]{grViz}} or the Graphviz code
if \code{gvcode = TRUE}.
}
\description{
Plot a diagram (directed graph) of top linkages
}
\section{Footnotes}{


* in the form of
\code{c('attribute_name1=attribute_value1', 'attribute_name2=attribute_value2, etc.')}

** \url{http://www.graphviz.org/doc/schema/attributes.xml}
}

