% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wiod.R
\name{sectors}
\alias{sectors}
\title{Extract sector codes and labels from a WIOD data.table}
\usage{
sectors(wiot)
}
\arguments{
\item{wiot}{WIOD data.table returned by \code{\link[wiod.diagrammer]{loadWIOD}}}
}
\value{
A \code{\link[data.table]{data.table}} with four columns:
\describe{
  \item{\code{SectorNr}}{WIOD numeric codes for industries and final use sectors (integer vector)}
  \item{\code{SectorLab}}{Full names of the WIOD sectors (character vector)}
  \item{\code{SectorCode}}{Short alphanumeric sector codes used in WIOD (character vector)}
  \item{\code{isFinal}}{Is a sector an industry (i.e. one of the 56 primary, secondary/manufacturing,
  or tertiary/services sectors that produce something) or a  final use sector
  (i.e. one of the 5 final consumption or investment sectors)? (logical vector)}
}
}
\description{
Extract sector codes and labels from a WIOD data.table
}

